Distance estimation
It works on floats, but It doesn't work on infinite precision, the derivative is too big and it overflows, I dame a hack that somewhat works for low zooms ( basically i calculate the derivative as double. Need to make infinite precision floating point numbering system for it to fully function.